# docs/openapi-3.0.yaml
openapi: 3.0.0
info:
  contact:
    email: support@volumeviz.io
    name: API Support
    url: https://github.com/mantonx/volumeviz/issues
  description: >-
    Docker volume monitoring API with comprehensive volume discovery, size
    calculation, and container attachment tracking. Focus on user-mounted
    volumes only.


    ## Volume-First Approach

    - Automatic discovery and filtering of user-mounted volumes

    - Excludes Docker infrastructure volumes (container filesystems, tmp volumes)

    - Real-time volume usage monitoring and size calculation

    - Container attachment tracking for each volume


    ## Features

    - Multi-method volume size calculation (du, find, stat)

    - Asynchronous scanning with progress tracking for large volumes

    - Real-time WebSocket progress updates during scan operations

    - High-performance caching with TTL-based invalidation

    - Comprehensive Prometheus metrics integration

    - Circuit breaker patterns for resilience


    ## WebSocket API

    WebSocket endpoint: `ws://localhost:8080/ws`


    ### Scan Progress Events

    Subscribe to real-time scan progress updates:

    ```json

    {

    "type": "subscribe",

    "event": "scan_progress",

    "data": {

    "volume_id": "volume_id_here",

    "scan_id": "scan_id_here"

    }

    }

    ```


    Progress updates are broadcast as:

    ```json

    {

    "type": "scan_progress_update",

    "data": {

    "scan_id": "scan_id",

...

...

...

    "phases": [

    {

...

...

...
paths:
...
  /api/v1/metadata/files/by-duration:
    get:
      description: Retrieve files filtered by video/audio duration with pagination support
      parameters:
        - description: Volume ID
          example: vol_123
          in: query
          name: volume_id
          required: true
          schema:
            type: string
        - description: Minimum duration in seconds
          example: 60
          in: query
          name: min_duration
          schema:
            type: integer
        - description: Maximum duration in seconds
          example: 3600
          in: query
          name: max_duration
          schema:
            type: integer
        - description: "Page number (default: 1)"
          example: 1
          in: query
          name: page
          schema:
            type: integer
        - description: "Items per page (default: 100, max: 500)"
          example: 100
          in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Files retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_api_v1_metadata.GetFilesByDurationResponse"
        "400":
          description: Invalid request parameters
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
      summary: Get files by duration
      tags:
        - metadata
  /api/v1/metadata/files/by-location:
    get:
      description: Retrieve files filtered by GPS coordinates with pagination support
      parameters:
        - description: Volume ID
          example: vol_123
          in: query
          name: volume_id
          required: true
          schema:
            type: string
        - description: GPS latitude
          example: 37.7749
          in: query
          name: latitude
          schema:
            type: number
            format: float64
        - description: GPS longitude
          example: -122.4194
          in: query
          name: longitude
          schema:
            type: number
            format: float64
        - description: Search radius in kilometers
          example: 10
          in: query
          name: radius
          schema:
            type: number
            format: float64
        - description: Filter files that have GPS data
          example: true
          in: query
          name: has_location
          schema:
            type: boolean
        - description: "Page number (default: 1)"
          example: 1
          in: query
          name: page
          schema:
            type: integer
        - description: "Items per page (default: 100, max: 500)"
          example: 100
          in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Files retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_api_v1_metadata.GetFilesByLocationResponse"
        "400":
          description: Invalid request parameters
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
      summary: Get files by location
      tags:
        - metadata
  /api/v1/metadata/files/by-media-kind:
    get:
      description: Retrieve files filtered by media kind (video, image, audio,
        document, etc.) with pagination
      parameters:
        - description: Volume ID
          example: vol_123
          in: query
          name: volume_id
          required: true
          schema:
            type: string
        - description: Media kind to filter by
          example: video
          in: query
          name: media_kind
          required: true
          schema:
            type: string
        - description: "Page number (default: 1)"
          example: 1
          in: query
          name: page
          schema:
            type: integer
        - description: "Items per page (default: 50, max: 200)"
          example: 50
          in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Files retrieved successfully
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
        "400":
          description: Invalid request parameters
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
        "404":
          description: Volume not found
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
      summary: Get files by media kind
      tags:
        - Metadata
  /api/v1/metadata/files/by-resolution:
    get:
      description: Retrieve files filtered by image/video resolution with pagination support
      parameters:
        - description: Volume ID
          example: vol_123
          in: query
          name: volume_id
          required: true
          schema:
            type: string
        - description: Image/video width
          example: 1920
          in: query
          name: width
          schema:
            type: integer
        - description: Image/video height
          example: 1080
          in: query
          name: height
          schema:
            type: integer
        - description: Minimum width
          example: 1024
          in: query
          name: min_width
          schema:
            type: integer
        - description: Maximum width
          example: 4096
          in: query
          name: max_width
          schema:
            type: integer
        - description: Minimum height
          example: 768
          in: query
          name: min_height
          schema:
            type: integer
        - description: Maximum height
          example: 2160
          in: query
          name: max_height
          schema:
            type: integer
        - description: "Page number (default: 1)"
          example: 1
          in: query
          name: page
          schema:
            type: integer
        - description: "Items per page (default: 100, max: 500)"
          example: 100
          in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Files retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_api_v1_metadata.GetFilesByResolutionRespon\
                  se"
        "400":
          description: Invalid request parameters
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
      summary: Get files by resolution
      tags:
        - metadata
  /api/v1/metadata/filters:
    get:
      description: Get available MIME types, media kinds, and extensions for filter
        dropdowns
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilterMetadataResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_mantonx_volumeviz_internal_models.ErrorR\
                  esponse"
      summary: Get filter metadata
      tags:
        - metadata
...
  "/files/{id}/details":
    get:
      description: Get comprehensive information about a specific file including metadata
      parameters:
        - description: File ID
          in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileDetailsResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_mantonx_volumeviz_internal_models.ErrorR\
                  esponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_mantonx_volumeviz_internal_models.ErrorR\
                  esponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_mantonx_volumeviz_internal_models.ErrorR\
                  esponse"
      summary: Get file details
      tags:
        - files
...
servers:
...
components:
  schemas:
...
    FilterMetadataResponse:
      properties:
        extensions:
          items:
            $ref: "#/components/schemas/ExtensionOption"
          type: array
        media_kinds:
          items:
            $ref: "#/components/schemas/MediaKindOption"
          type: array
        mime_types:
          items:
            $ref: "#/components/schemas/MimeTypeOption"
          type: array
      type: object
...
    internal_api_v1_metadata.FileDurationItem:
      properties:
        duration_seconds:
          type: integer
        id:
          type: integer
        media_type:
          type: string
        name:
          type: string
        path:
          type: string
        size:
          type: integer
      type: object
    internal_api_v1_metadata.FileLocationItem:
      properties:
        id:
          type: integer
        latitude:
          type: number
        longitude:
          type: number
        media_type:
          type: string
        name:
          type: string
        path:
          type: string
        size:
          type: integer
      type: object
...
    internal_api_v1_metadata.GetFilesByDurationResponse:
      properties:
        files:
          items:
            $ref: "#/components/schemas/internal_api_v1_metadata.FileDurationItem"
          type: array
        limit:
          type: integer
        page:
          type: integer
        total_count:
          type: integer
        total_pages:
          type: integer
      type: object
    internal_api_v1_metadata.GetFilesByLocationResponse:
      properties:
        files:
          items:
            $ref: "#/components/schemas/internal_api_v1_metadata.FileLocationItem"
          type: array
        limit:
          type: integer
        page:
          type: integer
        total_count:
          type: integer
        total_pages:
          type: integer
      type: object
...